requirements.txt
streamlit
langchain-core
langchain-text-splitters
langchain-openai
youtube-transcript-api
python-dotenv
openai